Crate kubert_prometheus_process

Process metrics for Prometheus.

This crate registers a collector that provides the standard set of Process metrics.

let mut prom = prometheus_client::registry::Registry::default();
if let Err(error) =
    kubert_prometheus_process::register(prom.sub_registry_with_prefix("process"))
{
    tracing::warn!(%error, "Failed to register process metrics");
}

Functions

register

Registers process metrics with the given registry. Note that the ‘process_’ prefix is NOT added and should be specified by the caller if desired.